+++ ascribe/pipeline/ox/src/io/eleven19/ascribe/pipeline/ox/package.scala
@@ -1,0 +1,5 @@
+package io.eleven19.ascribe.pipeline
+
+/** Re-export pipeline-core types at `io.eleven19.ascribe.pipeline.ox`. */
+package object ox:
+    export io.eleven19.ascribe.pipeline.core.{PipelineError, RewriteAction, RewriteRule}